UART read_out file
The script  is "plug-and-play". Whenever the RPi is powered up, the software is boot following the "rc.local" file. Within the "rc.local" file, the python script to read out the UART serial communication is started. 

The script starts by creating a data folder within "/home/pi/Data" with the UTC datetime in the name of the folder. Data is stored in memory, every 3 minutes a log file is created within the data folder.